fix(evaluation): stabilize curriculum diagnostics and report plots - Add final cleanup plan for the remaining architecture refactor phases. - Fix liftoff diagnostic task preparation for acceleration-constrained start-hold tasks. - Prefer duration extension over unsafe sample-rate changes when more diagnostic samples are needed. - Fall back to the original valid task with a warning if diagnostic task extension fails. - Keep full evaluation metrics and diagnostics based on all evaluation episodes. - Make main report plots use the same render rollout trace as the GIF when rendering is enabled. - Preserve full multi-episode failure reporting while keeping plots and GIF visually consistent. - Add plot trace metadata for scope, step count, termination and truncation state. - Add regression tests for liftoff diagnostics, multi-episode plot handling and render-rollout plot scope. - Validate with ruff, mypy, full pytest, manual curriculum training and final-stage benchmark evaluation.

+++ src/validation/validation_tasks.py
@@ -365,7 +365,7 @@ def _apply_task_start_hold(
     times = np.asarray(trajectory.times, dtype=float)
     positions = np.asarray(trajectory.positions, dtype=float)
     sample_interval = _sample_interval_sec(times)
-    hold_steps = max(1, int(round(requested_sec / sample_interval)))
+    hold_steps = max(1, round(requested_sec / sample_interval))
     effective_hold_sec = float(hold_steps * sample_interval)
     hold_times = times[0] + sample_interval * np.arange(hold_steps, dtype=float)
     hold_positions = np.repeat(positions[0].reshape(1, XYZ_DIMENSIONS), repeats=hold_steps, axis=0)
